Renamed the request cycle listeners and tests with longer names of coop and coep, removed CoopConfiguration and CoepConfiguration files
