+++ wicket-core/src/main/java/org/apache/wicket/protocol/http/WebApplication.java
@@ -37,10 +37,6 @@ import org.apache.wicket.WicketRuntimeException;
 import org.apache.wicket.ajax.AjaxRequestHandler;
 import org.apache.wicket.ajax.AjaxRequestTarget;
 import org.apache.wicket.ajax.AjaxRequestTargetListenerCollection;
-import org.apache.wicket.coep.CrossOriginEmbedderPolicyConfiguration;
-import org.apache.wicket.coep.CoepRequestCycleListener;
-import org.apache.wicket.coop.CrossOriginOpenerPolicyConfiguration;
-import org.apache.wicket.coop.CoopRequestCycleListener;
 import org.apache.wicket.core.request.mapper.IMapperContext;
 import org.apache.wicket.core.request.mapper.MountedMapper;
 import org.apache.wicket.core.request.mapper.PackageMapper;
